requirements.txt
Flask==2.3.3
Werkzeug==2.3.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.3
markupsafe==2.1.3
flask-login==0.6.2
flask-sqlalchemy==3.0.3
python-dotenv==1.0.0
azure-core==1.27.1
azure-search-documents==11.4.0
langchain==0.0.175
beautifulsoup4==4.12.2
chardet==5.1.0
lxml==4.9.2
nltk==3.8.1
tabulate==0.9.0
wrapt==1.14.1